requirements.txt
datasets==3.0.0
openai==1.55.3
pandas==2.2.3
peft==0.12.0
torch==2.4.1
transformers==4.44.2
xtuner==0.1.23
python-dotenv==1.0.1
wandb==0.17.5
sentence_transformers==3.4.1
tensorboard==2.19.0
anthropic==0.47.2